feat(markers): party marker type + 'markers in radius' for all markers
Replace the party-location panel/animations with a plain, deletable 'party' marker type and a generic in-radius tool available from the Marker Editor for every marker.
