+++ src/controllers/index.ts
@@ -32,6 +32,7 @@ export const Controllers = createRegistry({
   MarkersEditor: () => import("@/controllers/markers-editor").then(m => m.MarkersEditor),
   MarkersSettings: () => import("@/controllers/markers-settings").then(m => m.MarkersSettings),
   MarkerCreator: () => import("@/controllers/marker-creator").then(m => m.MarkerCreator),
+  MarkersInRadius: () => import("@/controllers/markers-in-radius").then(m => m.MarkersInRadius),
   MarkersOverview: () => import("@/controllers/markers-overview").then(m => m.MarkersOverview),
   MarketDealsOverview: () => import("@/controllers/market-deals-overview").then(m => m.MarketDealsOverview),
   MarketOverview: () => import("@/controllers/market-overview").then(m => m.MarketOverview),
@@ -41,7 +42,6 @@ export const Controllers = createRegistry({
   Minimap: () => import("@/controllers/minimap").then(m => m.Minimap),
   NamesbaseEditor: () => import("@/controllers/namesbase-editor").then(m => m.NamesbaseEditor),
   NotesEditor: () => import("@/controllers/notes-editor").then(m => m.NotesEditor),
-  PartyLocation: () => import("@/controllers/party-location").then(m => m.PartyLocation),
   ProductionChains: () => import("@/controllers/production-chains").then(m => m.ProductionChains),
   ProductionOverview: () => import("@/controllers/production-overview").then(m => m.ProductionOverview),
   ProvincesEditor: () => import("@/controllers/provinces-editor").then(m => m.ProvincesEditor),
