+++ src/index.html
@@ -2089,12 +2089,6 @@
               Namesbase
             </button>
             <button id="editNotesButton" data-tip="Click to open Notes Editor" data-shortcut="Shift + O">Notes</button>
-            <button
-              id="openPartyLocationButton"
-              data-tip="Click to open the Party Location panel (travel range: daily / weekly / monthly)"
-            >
-              Party
-            </button>
             <button id="editProvincesButton" data-tip="Click to open Provinces Editor" data-shortcut="Shift + P">
               Provinces
             </button>
