+++ src/services/io/load.ts
@@ -803,9 +803,6 @@ async function parseLoadedData(data: string[], mapVersion: string | null): Promi
       }
     }
 
-    // ensure the party marker exists (creates one for legacy maps that predate it)
-    Markers.ensurePartyLocation();
-
     // remove href from emblems, to trigger rendering on load
     select("#emblems").selectAll("use").attr("href", null);
     // draw data layers (not kept in svg)
