+++ src/services/io/save.ts
@@ -96,8 +96,8 @@ function prepareMapData(): string {
   if (cloneRuler) cloneRuler.innerHTML = ""; // always remove rulers
   const cloneTradeAnimation = cloneEl.querySelector("#tradeAnimation");
   if (cloneTradeAnimation) cloneTradeAnimation.innerHTML = ""; // always remove transient trade animations
-  const clonePartyRange = cloneEl.querySelector("#partyRange");
-  if (clonePartyRange) clonePartyRange.innerHTML = ""; // always remove the transient party travel-range ring
+  const cloneMarkerRadius = cloneEl.querySelector("#markerRadiusRing");
+  if (cloneMarkerRadius) cloneMarkerRadius.innerHTML = ""; // always remove the transient marker "in radius" ring
 
   const serializedSVG = new XMLSerializer().serializeToString(cloneEl);
